requirements.txt
pandas
numpy
openai
scikit-learn
matplotlib
python-dotenv